chore: PR文档更新 + 测试补充 + PyInstaller + ruff 修复 (92→39)
- docs: PR后续计划全部标记完成，新增已完成功能表格
- test: 新增 test_multifreq_dialog.py (3个) + test_quality_dialog.py (4个) 共7个测试
- build: echopype.spec 添加 multifreq_dialog/quality_dialog 到 hiddenimports
- style: ruff 修复 53 个问题（F821/F401/RUF046/RUF059/I001/PIE794/F811等）
- fix: workers.py 删除重复的 SingleTargetWorker 代码块

+++ src/core/region.py
@@ -163,7 +163,7 @@ def build_analysis_mask(
 
     # 排除表线以上
     if surface_sample is not None and not np.isnan(surface_sample):
-        surf_idx = int(round(surface_sample))
+        surf_idx = round(surface_sample)
         if surf_idx > 0:
             mask[:, :surf_idx] = False
